[VUFIND-1547] Standardize event namespace for search backend.
