+++ module/VuFind/src/VuFind/Search/BackendManager.php
@@ -161,7 +161,7 @@ class BackendManager
     {
         if (!$this->listeners->offsetExists($events)) {
             $listener = [$this, 'onResolve'];
-            $events->attach('VuFind\Search', Service::EVENT_RESOLVE, $listener);
+            $events->attach(Service::class, Service::EVENT_RESOLVE, $listener);
             $this->listeners->attach($events, $listener);
         }
     }
@@ -177,7 +177,7 @@ class BackendManager
     {
         if ($this->listeners->offsetExists($events)) {
             $listener = $this->listeners->offsetGet($events);
-            $events->detach($listener, 'VuFind\Search');
+            $events->detach($listener, Service::class);
             $this->listeners->detach($events);
         }
     }
