+++ module/VuFind/src/VuFind/Search/Factory/BlenderBackendFactory.php
@@ -130,12 +130,12 @@ class BlenderBackendFactory implements FactoryInterface
         $manager = $this->container->get('SharedEventManager');
 
         $manager->attach(
-            'VuFind\Search',
+            \VuFindSearch\Service::class,
             \VuFindSearch\Service::EVENT_PRE,
             [$backend, 'onSearchPre']
         );
         $manager->attach(
-            'VuFind\Search',
+            \VuFindSearch\Service::class,
             \VuFindSearch\Service::EVENT_POST,
             [$backend, 'onSearchPost']
         );
