+++ module/VuFind/src/VuFind/Search/Primo/InjectOnCampusListener.php
@@ -97,7 +97,7 @@ class InjectOnCampusListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_PRE,
             [$this, 'onSearchPre']
         );
