+++ module/VuFind/src/VuFind/Search/Solr/AbstractErrorListener.php
@@ -109,7 +109,7 @@ abstract class AbstractErrorListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_ERROR,
             [$this, 'onSearchError']
         );
