+++ module/VuFind/src/VuFind/Search/Solr/CustomFilterListener.php
@@ -106,7 +106,7 @@ class CustomFilterListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_PRE,
             [$this, 'onSearchPre']
         );
