+++ module/VuFind/src/VuFind/Search/Solr/FilterFieldConversionListener.php
@@ -75,7 +75,7 @@ class FilterFieldConversionListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_PRE,
             [$this, 'onSearchPre']
         );
