+++ module/VuFind/src/VuFind/Search/Solr/HierarchicalFacetListener.php
@@ -163,7 +163,7 @@ class HierarchicalFacetListener
         SharedEventManagerInterface $manager
     ) {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_POST,
             [$this, 'onSearchPost']
         );
