+++ module/VuFind/src/VuFind/Search/Solr/InjectConditionalFilterListener.php
@@ -86,7 +86,7 @@ class InjectConditionalFilterListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_PRE,
             [$this, 'onSearchPre']
         );
