+++ module/VuFind/src/VuFind/Search/Solr/MultiIndexListener.php
@@ -108,7 +108,7 @@ class MultiIndexListener
     public function attach(SharedEventManagerInterface $manager)
     {
         $manager->attach(
-            'VuFind\Search',
+            Service::class,
             Service::EVENT_PRE,
             [$this, 'onSearchPre']
         );
