+++ module/VuFindSearch/src/VuFindSearch/Service.php
@@ -32,7 +32,6 @@
 
 namespace VuFindSearch;
 
-use Laminas\EventManager\EventManager;
 use Laminas\EventManager\EventManagerInterface;
 use VuFindSearch\Backend\BackendInterface;
 use VuFindSearch\Backend\Exception\BackendException;
@@ -51,6 +50,8 @@ use VuFindSearch\Command\CommandInterface;
  */
 class Service
 {
+    use \VuFindSearch\Feature\EventManagerTrait;
+
     /**
      * Event identifiers.
      *
@@ -60,13 +61,6 @@ class Service
     public const EVENT_POST    = 'post';
     public const EVENT_ERROR   = 'error';
     public const EVENT_RESOLVE = 'resolve';
-
-    /**
-     * Event manager.
-     *
-     * @var EventManager
-     */
-    protected $events;
 
     /**
      * Cache resolved backends.
@@ -116,35 +110,6 @@ class Service
         $this->triggerPost($this, $args);
 
         return $command;
-    }
-
-    /**
-     * Set EventManager instance.
-     *
-     * @param EventManagerInterface $events Event manager
-     *
-     * @return void
-     * @todo   Deprecate `VuFind\Search' event namespace (2.2)
-     */
-    public function setEventManager(EventManagerInterface $events)
-    {
-        $events->setIdentifiers(['VuFind\Search', 'VuFindSearch']);
-        $this->events = $events;
-    }
-
-    /**
-     * Return EventManager instance.
-     *
-     * Lazy loads a new EventManager if none was set.
-     *
-     * @return EventManagerInterface
-     */
-    public function getEventManager()
-    {
-        if (!$this->events) {
-            $this->setEventManager(new EventManager());
-        }
-        return $this->events;
     }
 
     /**
